requirements.txt
transformers>=4.46.1
python-dotenv
google-generativeai

